Add new resource method: recordToToken
Create separate interfaces for AuthTokenRecord (how it's recorded in the database) and AuthToken (the thing we send in a response). Only `id` is the same between the two. `uid` gets replaced with the full `User` object; `refresh` is the Argon2 hash of the `refresh` saved to the database; and the `*_expiration` values are expanded into an `expiration` object to make them a little easier to work with. The `recordToToken` resource method takes an AuthTokenRecord object and transforms it into an AuthToken object.