Reorganize code
The `resource.ts` files in each collection was turning into a great big waste bin. JSON:API 1.1 has a lot of types, so it's useful to build up a library or transformer functions, but the answer isn't to include a junk drawer in each collection, it's to create a library of transformer functions. So that's what I did here. The `resource.ts` files are gone, and instead we've got a pretty solid start on our `utils/transformers` directory.