urlToFields
We used to have urlToUserFields, with the idea of having a separate one for each type, but as you get into relationships and compound documents, it's more often the case where we want all of the sparse fieldset data for all types at once. This stores that data in a Fields interface, and establishes a urlToFields method that can be expanded to handle additional types.